+++ integration_tests/test_open_exotic_dxf_files.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018-2019, Manfred Moitzi
+# Copyright (c) 2018-2020, Manfred Moitzi
 # License: MIT License
 import pytest
 import ezdxf
@@ -35,21 +35,23 @@ PROE_FILE = os.path.join(ezdxf.EZDXF_TEST_FILES, "ProE_AC1018.dxf")
 
 
 @pytest.mark.skipif(not os.path.exists(PROE_FILE),
-                    reason=f'File {PROE_FILE} not found')
-def test_read_ProE_file():
-    doc = ezdxf.readfile(PROE_FILE)
-    assert doc.filename is not None
-    assert doc.dxfversion is not None
+                    reason=f'File "{PROE_FILE}" not found')
+class TestProE:
+    @pytest.fixture
+    def filename(self):
+        return PROE_FILE
 
+    def test_read_ProE_file(self, filename):
+        doc = ezdxf.readfile(filename)
+        assert doc.filename is not None
+        assert doc.dxfversion is not None
 
-@pytest.mark.skipif(not os.path.exists(PROE_FILE),
-                    reason=f'File {PROE_FILE} not found')
-def test_recover_ProE_file():
-    doc, auditor = recover.readfile(PROE_FILE)
-    assert doc.filename is not None
-    assert doc.dxfversion is not None
-    assert len(auditor.errors) == 0
-    assert len(auditor.fixes) == 0
+    def test_recover_ProE_file(self, filename):
+        doc, auditor = recover.readfile(filename)
+        assert doc.filename == filename
+        assert doc.dxfversion is not None
+        assert len(auditor.errors) == 0
+        assert len(auditor.fixes) == 0
 
 
 FILE_CIVIL_3D = os.path.join(ezdxf.EZDXF_TEST_FILES,
@@ -57,21 +59,40 @@ FILE_CIVIL_3D = os.path.join(ezdxf.EZDXF_TEST_FILES,
 
 
 @pytest.mark.skipif(not os.path.exists(FILE_CIVIL_3D),
-                    reason=f'File {FILE_CIVIL_3D} not found')
-def test_read_civil_3d():
-    doc = ezdxf.readfile(FILE_CIVIL_3D)
-    assert doc.filename == FILE_CIVIL_3D
-    assert doc.dxfversion == 'AC1032'
+                    reason=f'File "{FILE_CIVIL_3D}" not found')
+class TestCorruptCivil3D:
+    """ The test file "AutodeskProducts/Civil3D_2018.dxf" contains binary data
+    in XRECORDS and is not valid for TrueView 2020!
 
+    """
+    @pytest.fixture
+    def filename(self):
+        return FILE_CIVIL_3D
 
-@pytest.mark.skipif(not os.path.exists(FILE_CIVIL_3D),
-                    reason=f'File {FILE_CIVIL_3D} not found')
-def test_recover_civil_3d():
-    doc, auditor = recover.readfile(FILE_CIVIL_3D)
-    assert doc.filename == FILE_CIVIL_3D
-    assert doc.dxfversion == 'AC1032'
-    assert auditor.has_errors is False
-    assert len(auditor.fixes) == 536
+    def test_readfile_and_preserve_errors_by_surrogate_escape(self, filename):
+        doc = ezdxf.readfile(filename, errors='surrogateescape')
+        assert doc.filename == filename
+        assert doc.dxfversion == 'AC1032'
+
+    def test_readfile_and_ignore_errors(self, filename):
+        doc = ezdxf.readfile(filename, errors='ignore')
+        assert doc.filename == filename
+        assert doc.dxfversion == 'AC1032'
+
+    def test_readfile_and_raise_exception(self, filename):
+        with pytest.raises(UnicodeDecodeError):
+            ezdxf.readfile(filename, errors='strict')
+
+        with pytest.raises(UnicodeDecodeError):
+            recover.readfile(filename, errors='strict')
+
+    def test_recover(self, filename):
+        doc, auditor = recover.readfile(filename, errors='ignore')
+        assert doc.filename == filename
+        assert doc.dxfversion == 'AC1032'
+        assert auditor.has_errors is True
+        assert len(auditor.errors) == 238, 'expected decoding errors'
+        assert len(auditor.fixes) == 536
 
 
 FILE_MAP_3D = os.path.join(ezdxf.EZDXF_TEST_FILES,
@@ -79,7 +100,7 @@ FILE_MAP_3D = os.path.join(ezdxf.EZDXF_TEST_FILES,
 
 
 @pytest.mark.skipif(not os.path.exists(FILE_MAP_3D),
-                    reason=f'File {FILE_MAP_3D} not found')
+                    reason=f'File "{FILE_MAP_3D}" not found')
 def test_read_map_3d():
     doc = ezdxf.readfile(FILE_MAP_3D)
     assert doc.filename == FILE_MAP_3D
@@ -90,7 +111,7 @@ GERBER_FILE = os.path.join(ezdxf.EZDXF_TEST_FILES, "Gerber.dxf")
 
 
 @pytest.mark.skipif(not os.path.exists(GERBER_FILE),
-                    reason=f'File {GERBER_FILE} not found')
+                    reason=f'File "{GERBER_FILE}" not found')
 def test_read_gerber_file():
     from ezdxf.lldxf.validator import is_dxf_file
     assert is_dxf_file(GERBER_FILE)